+++ src/promoted/code1_0022.c
@@ -263,7 +263,7 @@ extern s32 func_0010b300(s32 personaId);
 extern void func_0019ef30(u8 *unit, u16 action);
 extern void func_00106d40(s16 character, s16 slot, s16 value);
 extern u16 func_00104dc0(s16 character);
-extern u32 func_00104e30(s16 character);
+extern u16 func_00104e30(s16 character);
 extern void func_001056e0(s16 character, s16 value);
 extern void func_00105730(s16 character, s16 value);
 extern void func_00105d50(s16 character, u32 mask);
